#2020 removed duplicate metrics sending, renamed processor to have correct name
