+++ management/server/core/hub-manager/hub-manager-impl/src/main/java/io/subutai/core/hubmanager/impl/HubManagerImpl.java
@@ -62,8 +62,8 @@ import io.subutai.core.hubmanager.impl.processor.port_map.ContainerPortMapProces
 import io.subutai.core.hubmanager.impl.requestor.ContainerEventProcessor;
 import io.subutai.core.hubmanager.impl.requestor.ContainerMetricsProcessor;
 import io.subutai.core.hubmanager.impl.requestor.HubLoggerProcessor;
+import io.subutai.core.hubmanager.impl.requestor.P2pLogsSender;
 import io.subutai.core.hubmanager.impl.requestor.PeerMetricsProcessor;
-import io.subutai.core.hubmanager.impl.requestor.ResourceHostDataProcessor;
 import io.subutai.core.hubmanager.impl.requestor.VersionInfoProcessor;
 import io.subutai.core.hubmanager.impl.tunnel.TunnelEventProcessor;
 import io.subutai.core.hubmanager.impl.tunnel.TunnelProcessor;
@@ -84,7 +84,7 @@ import io.subutai.hub.share.json.JsonUtil;
 
 public class HubManagerImpl implements HubManager, HostListener
 {
-    private static final long RH_CONF_SEND_INTERVAL_SEC = TimeUnit.MINUTES.toSeconds( 15 );
+    private static final long P2P_LOG_SEND_INTERVAL_SEC = TimeUnit.MINUTES.toSeconds( 15 );
     private static final long METRICS_SEND_INTERVAL_SEC = TimeUnit.MINUTES.toSeconds( 10 );
     private static final int CONTAINER_METRIC_SEND_INTERVAL_MIN = 15;
 
@@ -92,8 +92,7 @@ public class HubManagerImpl implements HubManager, HostListener
 
     private final ScheduledExecutorService heartbeatExecutorService = Executors.newSingleThreadScheduledExecutor();
 
-    private final ScheduledExecutorService resourceHostConfExecutorService =
-            Executors.newSingleThreadScheduledExecutor();
+    private final ScheduledExecutorService peerLogsExecutor = Executors.newSingleThreadScheduledExecutor();
 
     private final ScheduledExecutorService resourceHostMonitorExecutorService =
             Executors.newSingleThreadScheduledExecutor();
@@ -135,7 +134,7 @@ public class HubManagerImpl implements HubManager, HostListener
 
     private HeartbeatProcessor heartbeatProcessor;
 
-    private ResourceHostDataProcessor resourceHostDataProcessor;
+    private P2pLogsSender p2pLogsSender;
 
     private ContainerEventProcessor containerEventProcessor;
 
@@ -191,16 +190,14 @@ public class HubManagerImpl implements HubManager, HostListener
     @Override
     public void onHeartbeat( final ResourceHostInfo resourceHostInfo, final Set<QuotaAlertValue> alerts )
     {
-        resourceHostDataProcessor.onHeartbeat( resourceHostInfo, alerts );
     }
 
 
     private void initHubRequesters()
     {
-        resourceHostDataProcessor = new ResourceHostDataProcessor( this, localPeer, monitor, restClient );
+        p2pLogsSender = new P2pLogsSender( this, localPeer, monitor, restClient );
 
-        resourceHostConfExecutorService
-                .scheduleWithFixedDelay( resourceHostDataProcessor, 20, RH_CONF_SEND_INTERVAL_SEC, TimeUnit.SECONDS );
+        peerLogsExecutor.scheduleWithFixedDelay( p2pLogsSender, 20, P2P_LOG_SEND_INTERVAL_SEC, TimeUnit.SECONDS );
 
         //***********
 
@@ -295,7 +292,7 @@ public class HubManagerImpl implements HubManager, HostListener
     @Override
     public void sendHeartbeat() throws HubManagerException
     {
-        resourceHostDataProcessor.process();
+        p2pLogsSender.process();
         heartbeatProcessor.sendHeartbeat( true );
         containerEventProcessor.process();
     }
@@ -325,13 +322,6 @@ public class HubManagerImpl implements HubManager, HostListener
     }
 
 
-    @Override
-    public void sendResourceHostInfo() throws HubManagerException
-    {
-        resourceHostDataProcessor.process();
-    }
-
-
     @RolesAllowed( { "Peer-Management|Delete", "Peer-Management|Update" } )
     @Override
     public void registerPeer( String email, String password, String peerName, String peerScope )
@@ -343,8 +333,6 @@ public class HubManagerImpl implements HubManager, HostListener
         RegistrationManager registrationManager = new RegistrationManager( this, configManager );
 
         registrationManager.registerPeer( email, password, peerName, peerScope );
-
-        sendResourceHostInfo();
 
         notifyRegistrationListeners();
 
@@ -714,7 +702,7 @@ public class HubManagerImpl implements HubManager, HostListener
     public void destroy()
     {
         heartbeatExecutorService.shutdown();
-        resourceHostConfExecutorService.shutdown();
+        peerLogsExecutor.shutdown();
         resourceHostMonitorExecutorService.shutdown();
     }
