+++ management/server/core/hub-manager/hub-manager-rest/src/main/java/io/subutai/core/hubmanager/rest/RestService.java
@@ -32,12 +32,6 @@ public interface RestService
     Response register( @FormParam( "email" ) String email, @FormParam( "password" ) String password,
                        @FormParam( "peerName" ) String peerName, @FormParam( "peerScope" ) String peerScope );
 
-    //Send resource host configurations
-    @POST
-    @Path( "/send-rh-configurations" )
-    @Produces( { MediaType.APPLICATION_JSON } )
-    Response sendRHConfigurations();
-
     //Register to HUB
     @GET
     @Path( "/dns" )
